Create UIChanges
These are the changes I made to streamline the user interface of the Bewander social networking website. I added a couple of labels in User Story 1 in order to clarify what the user was viewing on the site. I also added a navigation bar icon label ("Connect") which appears when the user hovers over the icon with their mouse. This involved adding HTML and editing the CSS file to display the "Connect" label only when the user hovers over the icon. It is hidden otherwise.